First transmission test
transmitting.ino writes either “a”, or “b” to serial, sending it out
the xrf

receiving.ino receives either “a”, or “b” and turns a light on or off
depending on what it receives.

When “b” is being sent both arduinos should be sending power out of
dpin 13.

a pot on spin 0 tells the transmitting arduino whether to transmit “a”
or “b”.